requirements.txt
bcbio-gff==0.6.4
biopython==1.68
cached_property==1.3.0
contextlib2==0.5.4
futures==3.0.5
pysam==0.10.0
shutilwhich==1.1.0
six==1.10.0
